Battle System Update
Party Pokemon Added To Player
When pokemon dies it automatically sends out the next pokemon
Added Random Wild Pokemon for Testing in the Level Manager
Added the ability for the player to hit the "Back" Button and go from the attack selection menu back to the action selection menu
added battle start observer and battle end observer.
Changed coding so the first pokemon in the party is sent out and the enemy pokemon sent out is the wild pokemon encountered
Created a game controller that controls functionality of the overworld and the battle system. prevents movement in the overworld, cleaned up animation for player