PartySystem Implimentation
Recreated, sliced up, labeled and organized the art for the party system
Created the corresponding prefab for each pokemon
Created the single party members code
Currently in the process of creating a static reference for each style of background. do not want to drag and drop on dynamic object.
Preset party system to be the same system from both the battle scene and the overworld. just different options available once a pokemon is selected