Party System
have the party screen functional
cut more of the sprite sheet up, more pokemon small sprites added in and tied to their pokemon base.
created buttons that the player can select through with no issue and the message board in the pokemon party does work
updates all the pokemon information upon opening including their sprites
cleaned up and polished a bit of the UI for party system.
created a signleton that holds the background of sprites for easier access instead of each having a reference of the gamobject being tied to it and switching between them.
Currently updates with the players pokemon party
No functionality added from the pokemon party when the pokemon are selected.