Gender Addition
Added in Genders to the inspector and added an editor function that when there is differences in the gender to pop up with more sprites for each style of the pokemon.
Entry hazards has been started.
Added in curse for the pokemon, one for the user where it damages itself and one for the target pokemon
Ghastly has been added in for additional testing, moves that do not effect certain types will not hurt the pokemon for further testing
2 rows of art has been chopped. new ways to implement this are being sought after