Battle system update Mar 20
cleaned up Curse for a volatile status
Added so multiple statuses can be implimented on the current pokemon, so both curse and confusion can be on the same pokemon without any issues, when curse is used on a pokemon the status does not stay on the user and be used continously to kill the owner.

if the pokemon dies before his turn ends and the opposing pokemon tries to attack while the check for battle over is false then it shall do the attack but there will be no target.

battle system cleaned up a bit and more implimentation and definittion as well as cleaning shit up