NPC AI developed
Created and implimented the NPC AI, they are moving and active
Can walk in specified path, at the start they shall detect if the path is walkable from the start.
updated entity and the game controller, created it so it shall run all entites through there.
Level design assiagns all of the entities through code
When a NPC is interacted with then they shall look at the target theyre talking to.
created that if the player is in their path they shall run towards the player and wait until the player moves, regardless if the player interrupts their path or not they shall not  break that path
Added in clown trainer sprite sheet, started prefab of trainer and clown adopts the sprite sheet without errors