Catching Mechanics added
Created all the pokeball mechanics, exactly like they had in the game
pokeball animations work but could be better.
Pokemon will be added to your party but not into the players PC as that has not been implimented.
items class created
pokeball class created that inherits from it
Pokeball catch rates added in, different pokeballs can be added in and implimented with different catch cates
all in game pokemon have had their catch rates added
Added an extension that when the dialog is waiting for an input to show a particular character
Item in battle system added in as one item for object pooling.
with trainer battles if the trainer is about to send out a new pokemon then it will ask to shift the pokemon.
Party system updated since it would clean up the pokemon and give them full health on each new battle.
Wild pokemon bug fixed where they wpuld have the same amount of health as the previous one.
Static Catching class added
Pokeballs added
Battle system bag set to throw pokeballs for the time being