Run From Battle Mechanic added in
Added in run mechanic exactly like how they have in the pokemon games.
cleaned up a bug that I implimentd in when moving around the code causing a null error.
Fixed issue where if there are more then 6 pokemon in the party it will cut the remainer out and state an error saying this Trainer has too many pokemon in his party.