Pokemon Levels
Experience implimented.
Added current pokemon experience yield
created the animation of the pokemon gaining experience in the HUD.
States on how much experience
pokemon will level up when enough experience is gained.
If the pokemon will level up then the bar will go to the position. upon gaining enough xp it will stop and the pokemon will level up.
pokemon will level up but the stat differences are not set. it only updates the UI.
pokemon base editor taken out and switched to it always being set in base since the cahnges were not saving.
if a pokemon learns a new move at set level with a free move slot then it will give the pokemon the move.
Created the manager/ prefab of where the pokemon chose a move to override but it has some bugs.
created prefab of buttons that go along with the level up manager
Cleaned up code in experience table
added in the symbol art for all the elemental types and added them in status condition singleton
sliced and added art in regards to elemental type and learning a new move