Finished Pokemon Learning Moves and Level Art
Finished pokemon learning new move implimentation.
added in some level art classes to compliment eachother
Pokemon background art was split up and added in
Sliced up a bunch of platforms that the pokemon stand on and recreated a bunch of the art
Added gender into the party
When in party it shows the correct pokemon sprite as well as the sprite given life and shows the animations.
When in battle and the player moves pokemon around it updates the pokemons order.
EV's finally finished. fully updated and implimented. pokemon will earn EV's but due to the pokemon not changing its stats yet they are not updated in yet.
fixed selector in the pokemon party.
Moved battle system awake functions to the game manager and it will call them from there.
party system was auto set off from the game manager script.
catching mechanic bug was corrected