Abilities Implimented
Created some basic abilities, have yet to try them out. they should work in theory but there is some testing to do regardless.
made the abilities database and organized it a bit due to the large amount of abilities that can be added. Added abiltities into the pokemon base stat, 2 starting ones and a hidden abilitiy for furute use. created a bad variation of passing it but it is to be expanded upon later.just to get things working