Abilities Added
added in the start of hyper cutter ability but did not place it in the battle system yet. Implimented all of the starting weather effects as abilities, tested them out and cleaned up the way weather effects work inside battle system so both the attacks and abilities work off the same method call.
Cleaned up abilities and preset it so it is easy to add more in via A-Z
Tested all abilities added in and no issues, working as intended