Leveling up Stat Changes on UI and with Pokemon
currently shows the stats of a pokemon leveling up, it will show the changes on the stats in a UI, it will have it all cleaned up and easy. pokemon will share experience now as well. other party pokemon sent out can get and recieve XP. not fully 100% implimented but it is working as needed. more testing will be done tmrw