Bug Fix, Clean Up, Small Additions
Added in Amnesia to help test
Cleaned up the Level Up UI, got rid of the messy testing area and cleaned it up to be nice and smooth. it now shows any pokemon that levels up with their stats and if the pokemon is not the one out in the field then it will show a little icon at the top right to signify which pokemon is receiveing the level up. Any new moves learned along with it work perfectly fine.
Battle unit cleaned up a bit, some of the calls that were void were changed into what theyre suppose to be as ienumerators. pokemon sprites were given their full animations. coroutines cleaned up. When the sprite reaches the main point it will play their starting animation. removed the coroutine that called itself at the end to move back.
Bugs cleaned out of entry hazard.Critical hits now negate any attack drop of the attacking pokemon and ignore any buffs on the defending pokemons defense.
Pokemon party. if the player starts the battle and the first pokemon in the party is dead it will reorder the party upon entry.
Found bug in Experience tables. clamped them just incase since some at level 1 would be in a negative value. Very slow had a bug due to it not being a float and the first 35 levels were 0 exp required.