Bug Removal and AI update
removed NPC states since is moving is essenatially sharing what state the entity is in.
Removed all of the bugs that were in the game and caught, gave pokemon their immunities.
Any secondary effects that it tried to infect the player with but would tell it that it doesnt effect them was removed.
Exp was fixed for all party pokemon.
Party order is now updated after the pokemon is dead at the starting position.
Immunities to particular status implimented.
Removed hit animation when pokemon was not effected by hit.
Fixed trainer bug where he was going to the trainer position when he seen him so it was creating errors.
changed AI movement, their paths wouldnt have an actual destination which would result in their movement pattern being broken