Battle System and Battle Unit Update
cleaned up the battle system start. Left it in a mess from the previous update.
Battle unit, cleaned up the start, changed the initial start of the battle and the way it went. Now everything goes through a different start method since there is an animation. plays out different for both wild pokemon and trainers but the battle stage is moves perfectly for both sides.  the original start is left if a pokemon is sent out as it does not move the stage. Tested and working perfectly