Bug Clean Up, Start Menu, Ability UI, Code Clean Up, Status Animations, Trainers animated, battle system starting animated
Cleaned up a series on bugs
If player lost all their pokemon would still be dead, healed upon losing
added a pokemon heal function within its own class script.
on end turn now needs the ID due to adding in animations so it will iterate between each one
Added in the start menu and implimented 2 of the buttons
Updated the party system to work with the Overworld Menu
game manager updated to work with the new start menu and the player controller
added animations for battle on start for when its with trainers and wild pokemon.
cleaned up battle system and split alot of the methods into their own regions.
When an ability is activated within certain conditions then it will display on the UI and show the ability that was activated
Big update from last 2 Days