Attacks, pokemon, Abilities additions and update
added in multiple strike pokemon moves,fixed variation and random number, as well as new moves ontop of it.
added in total amount of pokemon up to # 25 in the pokedex.
struggle implimented and working. added the ability for moves to be disabled but not implimented yet.
Static damage modifications added, specifically for false swipe but have it so more can be added in.
battle system updated for multimoves and new abiltites.
Ability database recieved a big update in abilities as well as alot of new functions