hold items base implimented in
Created the class on how the ability effects will be processed. currently throwing in standard items, havent tested them out but they are in the system. have to create the scriptable version of these items where their details shall be added in then just give their in battle effects be a reference through an ID. need to figure out a way that when it is used it gets removed and an animation is played