Inventory System Update
updated the inventory system so now the player can select a specific item and it will go into the next menu. it has 3 current options that show up with no functionality. currently only the cancel button works within it going back and forth. implimenting the party screen to interact with the inventory system seems like it will need work. going to change up a bit of the game manager, battle system, inventory system and the party system in the way they all react and refer to eachotehr into a smoother looking transition. going to change the in battle bool thats being passed and make it a static variable that can be privately set.
Currently a mess but just getting it working then will sift through everything and clean out all of the BS