Inventory Update
currently got the inventory working where the player can use the item and it will open up the bag. the logic works from in battle if the player opens up the bag then it will show the players inventory and if they cancel it shall go back to in the battle. have a bit of features implimented where then the player selects use then it shall show the pokemon party open up and they can select the item to use on the pokemon. but once an item is attempted to use it will just go back to the overworld. have to construct the full loop.