Dialog clean UP
before there were multiple areas that had repeated code, removed all different variations of dialog box and put it into one class. the dialog manager is what runs everything and the dialog box is just a reference to the positions in where the corresponding text is. this removes alot of duplicates but the next issue is creating a dialog box prefab. currently it is just different items but making the code work was the main issue