Code Refectoring
cleaned up a bit of the battle system, a lot of the actions and attacks, Removed some useless information, created prefabs of multiple objects and made everything alot smoother. Changed the main systems to be a derived class of core systems.  created a separate class to take everything in for the selection on the UI rather then reusing the same method in different classes