Invetnory Save Data and item saved for pokemon data
completely updated the save system to include all of the inventory items, its working perfectly, items have been updated to be saved on the pokemon. cleaned up the save data for the save class as well. updated more items into the game, alot of methods regarding items need to be implemented but still in additional phase