Hold Item Update & Critical Hit Ratio Fix
changed the critical hit ratio to a int, added it as a statboost due to some pokemon having specialized items and moves that can heighten the ratio on all critical hits. damage modfiers now carries the table and returns what the ratio shall be. added in alot more items into the game along with their functions and where they will go. cleaned up code as well.