TMHM created
created the HMHM script and updated the itembase to handle the specifics for it as well as the pokemon base class to hold the learnable TMHMs. currently the party screen will update saying that it will work or it will not.  updated and cleaned up the adjusting move priorities but it does not work still. will require further messing around with. HMHM half way implimented