TMHM fully implemented
Added in 5 first tms that were compatable from ultra series of 7th gen. tested and fully works correctly corresponding with the Learning UI that was made with the Battle System. pokemon that have the move already will show they do and cannot learn it, special dialog exists for that. pokemon that cannot have the item work on them will just show able or unable in the similar fashion in the game. tested out every scenario is working correctly