Updated Learning new moves UI
cleaned up the UI in the way it functions on the back end. Cleaned up the moves that were older and set in, fixed moves missing their descriptions and cleaned up a bit of work. found a bug in leech seed where the pokemon can die and it will heal them thinking that theyre inside of battle still and have a bug. Going to record a video and upload what I have done in the last little bit