cleaned up the extra canvas
moved battle system and party system over to the canvas rather then having its own. realized fade was the cause of the player not being able to click on the screen and removed the problem due to that. Going to add more context to the game and contiue growing it