Saving fully working, cuttable trees respawn
fully implimented and changed the saving, entities will save their position, direction and if they were battled perfectly. just need to add a camera bounds to the entities that can be seen for their positioning. items will be saved if theyre picked up. cuttable trees will save as implimented, only if it is in the current level