Advanced Saving System Implimented
fully converted the saving system to a serializable version. it will write everything to a txt file, before it would save via unitys player prefs. with the serialzable version it keeps it all in a class or struct. its alot cleaner then how it was before by splitting everything into a string and trying to convert it all over. it worked at the end of the day but the way this one is working is 100% better and to add onto this would not be hard at all