Abilities added New trainers Added
added in some pokemon trainers and updated brians party to what he wanted.  Big update for the pokemon trainers art and going to impliment what trainers are in the game as well as. starting to add in more abilities, Analytic aftermath are added as well as pressure. more to be added as time goes on