Level manager Update and pokemon trainers added
added in more art for pokemon, added in more trainers into the field and updated the wild pokemon so now there is just their pokemon base, range and % chance of finding it. in pokemon base i added in the wild items that the pokemon could possibly be holding. alot more pokemon trainers were added to the game. more level developement shall follow in time