NPC and item giver aspect added
Finished updated the ability for NPC's to give the player items. it will save if the player was given items. The NPC was updated with the newer AI but the way that the AI shall be developed will be elaborated on more.  More abilities will be added in the future