Ability Addition and bug fixed
battle system had a bug from when it was updated where the pokeball wasnt showing. Fixed that. changed damage details to split into defenders and attackers abilities. added more style of pokemon abilities into the game. 8 new abilties added in as well as some added in.
Pokemon reset needs to be looked at. could have created a potential bug when the pokemon is switched they heal.
Bug in summary fixed. it would show the pokemon last changes on the summary