cleaned up the Hold items functionality. Battle Effect items Added
removed the return as new derived class function that existed within these classes. any item that was removed off of it would be pulled off of it from the battle unit instead. cleaner way of functioning. When the item is removed its a function called off of the battle unit instead. Added in battle effect items, tested and working perfectly. more testing needs to be done for each of the berry items