Mist shield added, More moves Added
additional moves added and their specialties added for all moves from A-Cut except for the ones mentioned in notes and Z moves. along with the addition of mist which is a shield base move now and guard spec which does the same thing. added a movebase function that lowered the attack for crush grip and a move that will be added wring out. Bug with stat boost found and removed. it wouldnt show up on the moves in the inspector so it was adjusted