Entity AI updated, leanring new moves in UI, more moves added
Updated the learning new move UI to handle everything and removed it from battle and party. Cleaned up all the useless information from scriptable objects of the players name and their dialog to the actual script. so their art is the only thing leftover. added in a bunch of moves from D to near the end of F for physical moves. pokemon can now learn moves from evolutions and it will go through the iteration exactly like the game