Gift Pokemon and Attack Summary Updated
when selecting on the pokemon in the summary, when they go over the current list of moves it will show all of the pokemons moves. they cannot be switched around as of yet. Gift pokemon are now in the overworld and it works as intended as if it were from the original game