PC system update
added different pokemon into the box, having a test amount of pokemon. currently have a pointer that works the same as in game. you can pick up the pokemon and move it around to a different location and works perfectly. cannot close box with a pokmon in hand. the player can go in and out of the PC but it will duplicate pokemon right now. The pointer will work in all locations, if the player selects the top buttons it will show accordingly but the party button does not work. This is all of yesterdays work