PC update
completed multiple boxes, get the party aspect working, player can go into party and select the pokemon and move it. the party does not save. need to update the player party when exiting the box as well as making the pokemon move when one is taken out. they still sit in the same corresponding location. if pokemon moved from box one to 2 it will not have any issues. it shall store the information correctly