PC System Finshed
finally completed the pokemon system, have to comb through for any bugs, need to upgrade the system for the selections on moving, looking at summary, item and cancel.
Cleaned up all minor bugs, updated the save system to have all of the boxes saved to their individual aspects. each pokemon in it will save their own data as inteded in the game. minor bug on full screen when switching the pokemon box. bug clean ups and a video was saved and uploaded of all the new updates since the last video of pokemon