Captured Pokemon Mechanics
added the captured pokemon stuff in, if the party is full go to the pc and return what needs to be said.  If the pokemon goes to the pc it will find the first spot and be relocated there. tested out the saving and it works. Getting ready to start out the shop