Level Design additions & art additions
created some art assets that were loaded into the game, created more trainers for the player to interact with but never finalized the specifics on it. created 2/3 houses in the lakewood area (scenes to specific houses). cleaned up the way scenes load. changed a bit in the health bar but never tested it. made bug with the trainer controller where they detected the player and couldnt battle him, but was corrected