Pathfinding
created the grid pathfinding in a different project just incase of breaking this one. turns out it was for the better as some things were broken but they were not ported over. The pathfinding works exactly as intended. more tests need to be done but the same region that it worked with in test game. the NPC will walk around anywhere that is intended.

Added some more depth and cleaned up the limestone mountain area. made sure all the trainers that you battle will have a dialog. their current levels are not set