Level design, Move Additions and Bug Clean up
cleaned up the bug where when the player interacted with the NPC they would get stuck in a running animation. Fully developed all of the limestone mountain area. just need to add in all the new pokemon for the area. Move list was small but started adding in some status moves. will be adding more moves to give all the current pokemon a more polished feel.