Level Design, Pokemon Addition, Scripting
added in more pokemon to the game. Completely changed the pokemon evolution script and cleaned it up. tested out eveolving from leveling and giving an item and it works perfect. Added in more pokemon to the level manager for the limestone moutain. Added in specific 2nd gen pokemon for the area but didnt test them out yet.