Audio Manager implimented
cleaned up the confused animation as it was too long, have it working for when the pokemon receives it but not when they are going to attack.  cleaned up some trainer dialog. need to fix the sounds in the game atm, they will play really loud and when the battle is finished it should return the the current area music or the last clip played. Audio is working correctly