screenshots added in, level deign update
Added in Concordia Cave as well as a chunk of trainers within it. I did not develop the pokemon that could be found in the area as of yet. Im going to end up updating the pokemon API to fetch sprites from online, there is a chunk of pokemon that have their sprites created and will start to utilize that as well as the ? for pokemon who are null